Minor code re-org to rename 'Storage.h' to 'StorageSetup.h' to avoid collision with 'storage.h', on case-insensitive file systems.
